Avoid code duplication in interop for Numbers
